Organized Project Directory
Updated:
   -[App.js] (fixed import directories for App.js and all other files)
   -[scripts] (renamed to screens)
   -[style/styles] (renamed to styles/style)
 Added:
   -[frontend folder] (contains screens and styles folders)
   -[backend folder] (contatins Fire.js)
