+++ RoommatesPlus/frontend/screens/CreateHouseHold.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { styles } from '../style/styles';
+import { styles } from '../styles/style';
 import { Text, View, TextInput } from 'react-native';
 import { TouchableOpacity } from 'react-native-gesture-handler';
