+++ RoommatesPlus/frontend/screens/JoinHouseHold.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { styles } from '../style/styles';
+import { styles } from '../styles/style';
 import { Text, View, TextInput, TouchableOpacity, Button } from 'react-native';
 import {openDatabase} from 'react-native-sqlite-storage';
 import * as SQLite from 'expo-sqlite'
